refactor(Student): Create computer avatar folder and move files into it
