+++ src/assets/wise5/common/computer-avatar/ComputerAvatar.ts
