+++ src/assets/wise5/common/computer-avatar/ComputerAvatarSettings.ts
