+++ src/assets/wise5/common/computer-avatar/computer-avatar-component.ts
