+++ src/assets/wise5/components/dialogGuidance/DialogGuidanceContent.ts
@@ -1,5 +1,5 @@
 import { ComponentContent } from '../../common/ComponentContent';
-import { ComputerAvatarComponentContent } from '../../common/computer-avatar-component-content';
+import { ComputerAvatarComponentContent } from '../../common/computer-avatar/computer-avatar-component-content';
 import { FeedbackRule } from '../common/feedbackRule/FeedbackRule';
 
 export interface DialogGuidanceContent extends ComponentContent, ComputerAvatarComponentContent {
